+++ types/common/actions.ts
@@ -75,6 +75,7 @@ import type {
   ChangesetStatusChangedAction,
   ChangesetFileSetAction,
   ChangesetFileRemovedAction,
+  ChangesetFilesReviewedChangedAction,
   ChangesetContentChangedAction,
   ChangesetOperationsChangedAction,
   ChangesetOperationStatusChangedAction,
@@ -174,6 +175,7 @@ export const enum ActionType {
   ChangesetStatusChanged = 'changeset/statusChanged',
   ChangesetFileSet = 'changeset/fileSet',
   ChangesetFileRemoved = 'changeset/fileRemoved',
+  ChangesetFilesReviewedChanged = 'changeset/filesReviewedChanged',
   ChangesetContentChanged = 'changeset/contentChanged',
   ChangesetOperationsChanged = 'changeset/operationsChanged',
   ChangesetOperationStatusChanged = 'changeset/operationStatusChanged',
@@ -293,6 +295,7 @@ export type StateAction =
   | ChangesetStatusChangedAction
   | ChangesetFileSetAction
   | ChangesetFileRemovedAction
+  | ChangesetFilesReviewedChangedAction
   | ChangesetContentChangedAction
   | ChangesetOperationsChangedAction
   | ChangesetOperationStatusChangedAction